Refactor: 美容室の予約ページ構造の変更
